producer/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.117.1
fonttools==4.60.0
idna==3.10
joblib==1.5.2
kafka-python==2.2.15
kiwisolver==1.4.9
lz4==4.3.3
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
# Core dependencies
kafka-python==2.2.15
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy

# Dashboard dependencies
streamlit
plotly
redis
altair

# Web framework dependencies
fastapi
pydantic
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
